fix: refactor of how images and Image jsx elements are loaded and handled
