+++ components/images.js
@@ -1,67 +1,52 @@
-export const actionBar = require('../assets/images/Action_Bar.png')
-export const accounts = require('../assets/images/account.png')
-export const adaptiveIcon = require('../assets/images/adaptive-icon.png')
-export const alarm1 = require('../assets/images/Alarm-1.png')
-export const alarm = require('../assets/images/Alarm.png')
-export const appIcon = require('../assets/images/appicon.png')
-export const clockFaceHour = require('../assets/images/Clock_face_hour.png')
-export const favIcon = require('../assets/images/favicon.png')
-export const footer = require('../assets/images/Footer.png')
-export const group2 = require('../assets/images/Group2.png')
-export const group3 = require('../assets/images/Group3.png')
-export const group4 = require('../assets/images/Group4.png')
-export const keyIcon = require('../assets/images/KeyIcon.png')
-export const more = require('../assets/images/More.png')
-export const NY = require('../assets/images/NY.png')
-export const PeriodSelector = require('../assets/images/PeriodSelector.png')
-export const puzzle = require('../assets/images/Puzzle.png')
-export const rectangle1 = require('../assets/images/Rectangle1.png')
-export const rectangle8 = require('../assets/images/Rectangle8.png')
-export const rectangle16 = require('../assets/images/Rectangle16.png')
-export const rectangle18 = require('../assets/images/Rectangle18.png')
-export const rectangle20 = require('../assets/images/Rectangle20.png')
-export const secondaryAction = require('../assets/images/secondaryAction.png')
-export const seperator = require('../assets/images/Separator.png')
-export const singleLineItem = require('../assets/images/SingleLineItem.png')
-export const splash = require('../assets/splash.png')
-export const timeSelector = require('../assets/images/TimeSelector.png')
-export const settings = require('../assets/images/settings.png')
-export const stats = require('../assets/images/stats.png')
-export const alarmHeader = require('../assets/images/alarmHeader.png')
-export const clear = require('../assets/images/clear.png')
+import { Image } from 'react-native'
 
-export default {
+export const images = {
+  actionBar: require('../assets/images/Action_Bar.png'),
+  accounts: require('../assets/images/account.png'),
+  adaptiveIcon: require('../assets/images/adaptive-icon.png'),
+  alarm1: require('../assets/images/Alarm-1.png'),
+  alarm: require('../assets/images/Alarm.png'),
+  appIcon: require('../assets/images/appicon.png'),
+  clockFaceHour: require('../assets/images/Clock_face_hour.png'),
+  favIcon: require('../assets/images/favicon.png'),
+  footer: require('../assets/images/Footer.png'),
+  group2: require('../assets/images/Group2.png'),
+  group3: require('../assets/images/Group3.png'),
+  group4: require('../assets/images/Group4.png'),
+  keyIcon: require('../assets/images/KeyIcon.png'),
+  more: require('../assets/images/More.png'),
+  nY: require('../assets/images/NY.png'),
+  periodSelector: require('../assets/images/PeriodSelector.png'),
+  puzzle: require('../assets/images/Puzzle.png'),
+  rectangle1: require('../assets/images/Rectangle1.png'),
+  rectangle8: require('../assets/images/Rectangle8.png'),
+  rectangle16: require('../assets/images/Rectangle16.png'),
+  rectangle18: require('../assets/images/Rectangle18.png'),
+  rectangle20: require('../assets/images/Rectangle20.png'),
+  secondaryAction: require('../assets/images/secondaryAction.png'),
+  seperator: require('../assets/images/Separator.png'),
+  singleLineItem: require('../assets/images/SingleLineItem.png'),
+  splash: require('../assets/splash.png'),
+  timeSelector: require('../assets/images/TimeSelector.png'),
+  settings: require('../assets/images/settings.png'),
+  stats: require('../assets/images/stats.png'),
+  alarmHeader: require('../assets/images/alarmHeader.png'),
+  clear: require('../assets/images/clear.png')
+}
 
-  actionBar,
-  accounts,
-  adaptiveIcon,
-  alarm1,
-  alarm,
-  appIcon,
-  clockFaceHour,
-  favIcon,
-  footer,
-  group2,
-  group3,
-  group4,
-  keyIcon,
-  more,
-  NY,
-  PeriodSelector,
-  puzzle,
-  rectangle1,
-  rectangle8,
-  rectangle16,
-  rectangle18,
-  rectangle20,
-  secondaryAction,
-  seperator,
-  singleLineItem,
-  splash,
-  timeSelector,
-  settings,
-  stats,
-  alarmHeader,
-  clear
-
+// Write images function that returns a resized image.
+export function getImage(image, width, height, customStyle, resizeMode = 'contain') {
+  return (
+    <Image
+      source = {image}
+      resizeMode = {resizeMode}
+      style = {{
+        ...{
+          width: width,
+          height: height
+        },
+        ...customStyle
+      }}
+    />
+  )
 }
